requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
scikit-learn>=1.5.0
pandas>=2.1.3
numpy>=1.26.0
joblib>=1.3.2
python-multipart>=0.0.6
requests>=2.31.0
python-jose>=3.3.0
cryptography>=41.0.0
passlib>=1.7.4
bcrypt>=4.0.0
python-dotenv>=1.0.0
